Rename utils to translator to match other plugins
