+++ mopidy_gmusic/library.py
@@ -7,8 +7,8 @@ from cachetools import LRUCache
 from mopidy import backend
 from mopidy.models import Album, Artist, Ref, SearchResult, Track
 
-from mopidy_gmusic.utils import album_to_ref, artist_to_ref, track_to_ref
-from mopidy_gmusic.utils import create_id, get_images
+from mopidy_gmusic.translator import album_to_ref, artist_to_ref, track_to_ref
+from mopidy_gmusic.translator import create_id, get_images
 
 logger = logging.getLogger(__name__)
